Moved netlist info
